+++ packages/algob/src/builtin-tasks/sign-multisig.ts
@@ -1,3 +1,4 @@
+import { getPathFromDirRecursive } from "@algo-builder/runtime";
 import path from "path";
 
 import { task } from "../internal/core/config/config-env";
@@ -25,13 +26,14 @@ async function multiSignTx (
     return;
   }
   const rawTxn = loadSignedTxnFromFile(taskArgs.file);
+  const sourceFilePath = getPathFromDirRecursive(ASSETS_DIR, taskArgs.file) as string;
   if (rawTxn === undefined) {
     console.error("Error loading transaction from the file.");
     return;
   }
   const signedTxn = signMultiSig(signerAccount, rawTxn);
   const outFileName = taskArgs.out ?? taskArgs.file.split(".")[0] + "_out.txn";
-  const outFilePath = path.join(ASSETS_DIR, outFileName);
+  const outFilePath = path.join(path.dirname(sourceFilePath), outFileName);
   await writeToFile(signedTxn.blob, taskArgs.force, outFilePath);
 }