+++ packages/algob/src/index.ts
@@ -1,3 +1,5 @@
+import { addressToPk, stringToBytes, uint64ToBigEndian } from "@algo-builder/runtime";
+
 import * as ERRORS from "./errors/errors-list";
 import {
   createMsigAddress,
@@ -15,7 +17,7 @@ import { balanceOf, printAssets, printGlobalStateSSC, printLocalStateSSC, readGl
 import { executeSignedTxnFromFile, executeTransaction, getSuggestedParams, mkTxParams } from "./lib/tx";
 import * as runtime from "./runtime";
 import * as types from "./types";
-
+const convert = { uint64ToBigEndian, addressToPk, stringToBytes };
 export {
   ERRORS,
   mkAccounts,
@@ -40,5 +42,6 @@ export {
   types,
   signMultiSig,
   signLogicSigMultiSig,
-  runtime
+  runtime,
+  convert
 };