+++ packages/runtime/src/index.ts
@@ -4,7 +4,7 @@ import { applyErrorMessageTemplate } from "./errors/runtime-errors";
 import { parseZodError } from "./errors/validation-errors";
 import { Interpreter } from "./interpreter/interpreter";
 import { loadASAFile, overrideASADef, validateASADefs } from "./lib/asa";
-import { loadFromYamlFileSilent, loadFromYamlFileSilentWithMessage } from "./lib/files";
+import { getPathFromDirRecursive, loadFromYamlFileSilent, loadFromYamlFileSilentWithMessage, lsTreeWalk } from "./lib/files";
 import { addressToPk, parseSSCAppArgs, stringToBytes, uint64ToBigEndian } from "./lib/parsing";
 import { encodeNote, getFromAddress, mkTransaction } from "./lib/txn";
 import { parser } from "./parser/parser";
@@ -31,5 +31,7 @@ export {
   validateASADefs,
   overrideASADef,
   parser,
+  lsTreeWalk,
+  getPathFromDirRecursive,
   types
 };