+++ packages/runtime/src/parser/parser.ts
@@ -4,14 +4,14 @@ import { Interpreter } from "../interpreter/interpreter";
 import {
   Add, Addr, Addw, And, AppGlobalDel, AppGlobalGet, AppGlobalGetEx,
   AppGlobalPut, AppLocalDel, AppLocalGet, AppLocalGetEx, AppLocalPut,
-  AppOptedIn, Arg, Balance, BitwiseAnd, BitwiseNot, BitwiseOr,
+  AppOptedIn, Arg, Assert, Balance, BitwiseAnd, BitwiseNot, BitwiseOr,
   BitwiseXor, Branch, BranchIfNotZero, BranchIfZero, Btoi,
   Byte, Bytec, Bytecblock, Concat, Div, Dup, Dup2, Ed25519verify,
   EqualTo, Err, GetAssetDef, GetAssetHolding, Global, GreaterThan,
   GreaterThanEqualTo, Gtxn, Gtxna, Int, Intc, Intcblock, Itob,
   Keccak256, Label, Len, LessThan, LessThanEqualTo, Load, Mod,
-  Mul, Mulw, Not, NotEqualTo, Or, Pop, Pragma, Return, Sha256,
-  Sha512_256, Store, Sub, Substring, Substring3, Txn, Txna
+  Mul, Mulw, Not, NotEqualTo, Or, Pop, Pragma, PushBytes, PushInt, Return, Sha256,
+  Sha512_256, Store, Sub, Substring, Substring3, Swap, Txn, Txna
 } from "../interpreter/opcode-list";
 import { LogicSigMaxCost, LogicSigMaxSize, MaxAppProgramCost, MaxAppProgramLen, OpGasCost } from "../lib/constants";
 import { assertLen } from "../lib/parsing";
@@ -124,12 +124,17 @@ opCodeMap[2] = {
  * TEALv3 opcodes: https://developer.algorand.org/articles/introducing-teal-version-3/
  */
 opCodeMap[3] = {
-  ...opCodeMap[2]
+  ...opCodeMap[2],
+
+  assert: Assert,
+  swap: Swap,
+
+  // optimized opcodes for pushing uint64s and byte slices to the stack
+  pushint: PushInt,
+  pushbytes: PushBytes
 
   /*
-  assert: Assert,
   dig: Dig,
-  swap: Swap,
   select: Select,
 
   // txn ops in tealv3
@@ -141,10 +146,6 @@ opCodeMap[3] = {
   setbit: Setbit,
   getbyte: Getbyte,
   setbyte: Setbyte,
-
-  // optimized opcodes for pushing uint64s and byte slices to the stack
-  pushint: Pushint,
-  pushbytes: Pushbytes,
 
   // stateful op (mode = application)
   min_balance: MinBalance